combinations python code
program is working however there is a problem with the sum and the multiply feature.

if you are looking for a range in the sum or the multiply feature, it needs to run twice and then needs to get common parts of 2 results. but it combines 2 lists return a combined list.

How we can solve this?